Keep text_width in config
Moving text_width to utils forced config to import it back (via a lazy
import inside box_label_width) to keep DRAWING_ATTRIBUTES self-contained,
since utils already imports config. Keeping text_width and its constants
(POINTS_PER_INCH, TEXT_WIDTH_RESOLUTION) in config avoids that round-trip
and lets box_label_width call it directly.
